Refactor SEG-Y Rev1 and Rev0 standards, move core descriptions to Rev0
Refactored the code to better manage and organize the SEG-Y Rev1 and Rev0 standards by creating a new file for Rev0. A substantial amount of the descriptions for SEG-Y standards were moved into this new Rev0 file, now imported into Rev1. This change simplifies the Rev1 and allows for clearer distinction and better organization between the two revisions.