Merged PR 24523: updated get_spec function to return copy of requested spec
updated get_spec function to return copy of requested spec.
The reason for this is to ensure that a new copy of the "spec" object is returned on each call to this function, instead of returning a reference to the same object. In the case where the same reference is returned, any modifications made to the spec will affect all other objects using the same spec name/revision.