Add Code of Conduct file
A new "Code of Conduct" file has been created and added to the project. This file outlines our community's pledge, standards, and enforcement responsibilities. It defines acceptable and unacceptable behaviors, and specifies the consequences for violating these standards. This step is taken to ensure our community remains open, diverse, inclusive, and welcoming.