Add CONTRIBUTING.md guide
A contributing guide has been added to the project to provide contributors with necessary information and resources. It includes instructions on how to report bugs, request features, set up a development environment, install and run the project, test the project, and submit changes. It also links to important resources like source code, documentation, issue tracker, and code of conduct.