Enhance Documentation (#28)
* Refactor aliased binary/trace header descriptors to be StructuredDataTypeDescriptor

Removed specific BinaryHeaderDescriptor and TraceHeaderDescriptor classes, instead utilizing StructuredDataTypeDescriptor. This change simplifies the codebase by reducing redundancy and incorporating a more general class where specific ones were previously used. This refactoring also involved updating tests and other parts where the removed classes were implemented.

* Update docstring for SEG-Y descriptor customization

The commit enhances the docstring for the function that customizes an existing SEG-Y descriptor. The new docstring includes details about the arguments it takes and the return value, providing clearer information for developers on how to use this function.

* Rework documentation structure in data models section

The documentation skeleton has been revamped for better clarity and understanding. `header.md` has been deleted, `data_type.md` is newly added, and minor edits in `trace.md` and `index.md` files. The changes are aimed to better illustrate the core concepts and structures in data models.

* Add SEG-Y descriptor documentation

This commit introduces documentation for the SEG-Y file structure to the data models sections. It provides a comprehensive overview and details about various components of the SEG-Y file format, such as SEGY-standard, text and binary file headers, trace and extended text headers. The information outlined can be beneficial for users for a better understanding of the SEG-Y seismic data files in terms of both reading and processing them programmatically.

* Update docstrings in data_type.py and add examples

Improved clarity and detail of docstrings in data_type.py. The updates include changes to descriptions for enumeration and descriptor classes. Additionally, example usage has also been incorporated for these classes, providing valuable context and clarification for developers on how these functions operate.

* Update documentation references and wording

The commit modifies two documentation strings. In `data_type.py`, it updates references to class components to improve linkage. Meanwhile, in `segy.py`, it rewords the description of SegyStandard for clarity.

* Update byte range notation in documentation

The commit updates the byte range notation for different fields in the source code documentation. Changes are purely stylistic, transitioning from using a '-' delimiter to using detailed sentences for clarity. The byte ranges themselves are unchanged.

* Update SEG-Y file class and methods documentation

Enhanced the class and method docstrings in src/segy/file.py to provide a more detailed description of the module's logic. This includes information on how to open a SEG-Y file based on a custom spec as well as the purpose of the SegyFile class and its various accessors. Additional comments also address a potential issue for further investigation.

* Update pydantic-settings autodoc settings in conf.py

The autodoc settings for pydantic models and settings have been updated in the conf.py file. This includes changes to member ordering and the visibility of summaries for fields and validators.

* Update type for 'REVISION' in SegyFileSettings

In the 'SegyFileSettings' class, the 'REVISION' field type was updated to accept 'int', 'float', or 'None' types. Additionally, the instantiation of 'BINARY' was modified utilizing 'default_factory' inside the 'Field' instance.

* Update API reference with settings

The API reference documentation has been updated with new sections for more clarity. Changes include an enriched explanation of the SEG-Y File class, along with added details for various configurations like SegyFileSettings, SegyBinaryHeaderSettings, ExtendedTextHeaderSetting, SampleIntervalSetting, and SamplesPerTraceSetting.

* Add settings documentation

A new documentation page about settings management was added. It provides details about how to initialize and handle instances of SegyFileSettings. It further explains how environment variables can be used to configure SegaFileSettings.

* Remove line continuation character in installation.md

The '\' line continuation character was removed from the text in `docs/installation.md`. This simplifies the readability of the document and prevents any potential rendering issues.

* Fix typos and add pip and pypi links in README

The commit corrects a typo in our project README file, tidies up the format and also adds direct pip and PyPi links. This enhancement is intended to facilitate easier access to Python packages and therefore improve the user's experience.

* Add x-ref to useful classes.

* Remove empty link from settings.md documentation file

* Update documentation for USE_PANDAS setting

Clarified the USE_PANDAS setting in the documentation. Added information about its non-applicability to trace data, explaining that trace data is always returned as Numpy arrays. Also mentioned that the option to use Numpy for headers is currently disabled but will be available in a future release.

* Update trace documentation in data models

The documentation for seismic traces in data models has been significantly expanded. The new documentation includes comprehensive definitions and explanations, as well as examples for defining a trace, trace header descriptor, and trace data descriptor. It also introduces how to create a trace descriptor both programmatically and from JSON.

* Remove SampleIntervalSetting from api_reference and improve trace documentation links

SampleIntervalSetting was removed from the API reference doc. The trace documentation has been improved with precise links, replacing the Trace Header and Trace Data link references with trace-header-descriptor and trace-data-descriptor. Duplicate SegyFileSettings link reference from settings file deleted.

---------

Co-authored-by: Altay Sansal <altay.sansal@tgs.com>