Generalize transforms and headers (#50)
* Refactor header and transformation handling code

The previously contained header and transformation handling functionalities in 'header.py' are split into 'arrays.py' and 'transforms.py'. This separation of concerns provides a more understandable structure and improves maintainability of the code. Additionally, 'arrays.py' is enriched by subclassing ndarray from Numpy for a better user experience.

* Refactor transformation tests into new file

The test_header_accessor.py file was becoming cluttered with transformation tests, so these have been extracted into their own file, test_transforms.py for better organization. Additional tests for the TransformStrategyFactory, scale, and scale_field have been added to this new file.

* Simplify assertions in SEGY file tests bc of new array logic

* Update type bound in TransformStrategy

The type bound for `T` in `TransformStrategy` has been changed from `BaseNDArray` to `NDArray`. This change aligns with the removal of dependency on `BaseNDArray` in this file, and ensures type consistency throughout the codebase.

* Update type bound in TransformStrategy

The type bound for `T` in `TransformStrategy` has been changed from `BaseNDArray` to `NDArray`. This change aligns with the removal of dependency on `BaseNDArray` in this file, and ensures type consistency throughout the codebase.

* Refactor code for binary, trace header access and transformations

The code has been refactored to improve binary header handling. HeaderAccessor has been replaced with HeaderNDArray for overall consistency. Instead of using the queue_transform method, a new TransformPipeline is created and transformations are added to it. This makes the transformation process clearer and more flexible.

* Add NumPy subclassing details to comments

In the arrays.py file, references and details for NumPy array subclassing were added. The added information includes a link to the official NumPy documentation about array subclassing basics. This provides clarity on the implementation for future developers.

* Move apply_transforms field to SegyFileSettings

The apply_transforms field was removed from the general settings in config.py and added to the SegyFileSettings class. This change aims to ensure that transformations are applied specifically in the context of file settings, improving maintainability and readability.

* Refactor SEGY file handling and transformation code

This commit refactors the SEGY file handling and transformation code for improved readability and consistency. Notably, it simplifies the transformation pipeline by directly returning transformed binary headers and data. It also changes the type hints in various areas to be more accurate.

* Add exception handling tests for transform strategies

In this commit, multiple tests were added to ensure proper exception handling in the transformation strategies. Specifically, tests were added to validate the behavior when a non-struct is given to the scale field strategy and when an unknown transformation type is requested from the factory. The exception type in the factory was also updated from `ValueError` to `KeyError`.

* Optimize scaling operations in transforms.py

The previous implementation utilized list comprehensions to scale data. This update changes that and directly applies the scale factor to the data, achieving the same result but with more efficient code. This change enhances performance and maintains the same functionality for the transform and inverse_transform methods.

---------

Co-authored-by: Altay Sansal <altay.sansal@tgs.com>