Refactor endianness handling in data types
The main change in this commit is the restructuring of how endianness handling is done in data type definitions. The responsibility for setting the endianness is moved out from the ScalarType and DataTypeDescriptor and into the StructuredFieldDescriptor. This change simplifies the datatype handling in ScalarType and DataTypeDescriptor, and delegates the endianness handling to a higher level in the data structure hierarchy.