Refactor endianness handling in trace schema
The changes involve removing the endianness field from the Trace Data descriptor class and adding it to the Trace descriptor class. This allows endianness handling at the level of the entire trace, including headers and data, instead of just at the data samples level. The dtype property computations are also adjusted accordingly to consider the new endianness placement.