Adjust SEGY revision setting in factory
The code has been modified to only set the SEGY revision value if it is not equal to REV0. This avoids unnecessary operations and grants more efficiency when handling SEGY standards, primarily when the REV0 standard is in use. This change is applied within the binary header configuration of the factory module in SEGY.