Add tests for TraceDescriptor and refactor TextHeaderDescriptor tests
The changes introduced new tests that verify the functionality of the TraceDescriptor. This was added within a new Python file named "test_schema_trace.py". It also refactored existing tests for TextHeaderDescriptor, simplifying the code and reducing the number of total lines. Furthermore, the file "test_header.py" was renamed to "test_schema_header.py" to better reflect its contents.