Override Numpy's copy method to `SegyArray` to ensure uninterpreted bytes are preserved. (#78)
* Add special copy method to SegyArray class

A 'copy' method, which maintains raw binary data and padding bytes, has been added to the SegyArray class in arrays.py. This is essential for SEG-Y data processing where unparsed fields are present but preserving raw binary data integrity is crucial.

* Update literal values and copy method static type checking

* Simplify array copy process

* make mypy happy

* Add test for SegyArray copy method

The test has been added to confirm that the copy method of SegyArray can successfully copy a segy array with the exact underlying buffer. The test also verifies that the memory between the original segy array and its copy are not shared.

---------

Co-authored-by: Altay Sansal <altay.sansal@tgs.com>