Headers: Change interface to use byte location instead of offset (#117)
* Update SEGY Schema to use byte locations instead of offsets

Updated the fields in the SEGY schema data_type.py file to use byte locations instead of offsets. The switch from offsets (0-based) to the byte locations (1-based) should make indexing more intuitive, especially for those familiar with the SEGY file standard. Changes were also propagated to associated test files.

* Update byte offset in trace data model documentation

The document "trace.md" in the data model documents has been updated to reflect changes in the byte offset for inline and crossline fields. The changes were from offset 188 to byte 189 for the inline field, and offset 192 to byte 193 for the crossline field.

* update lockfile

---------

Co-authored-by: Altay Sansal <altay.sansal@tgs.com>