Interpret SEG-Y revision in binary header (#190)
* Add SEG-Y revision interpretation transform

Introduced a new `SegyRevisionTransform` class to interpret the SEG-Y revision field in the binary header. This transform has been added to the `transform_map` and applied in the SEG-Y file processing pipeline. The code raises a `NotImplementedError` for SEG-Y revisions 2 and above.

* Prevent transformation for SEGY revision zero

Add a check to exit the transform method early if the SEGY revision is zero. This avoids unnecessary processing and potential errors when the revision is not set or is invalid.

* Fix segy_revision dtype check in transform method

Updated the transform method to check for the presence of the 'segy_revision' field in the data's dtype. This ensures compatibility with data structures that may not contain the 'segy_revision' key.

* Remove NotImplementedError for SegyStandard Rev2+

This change eliminates the exception raised for SEGY revision 2 or higher, allowing the code to continue without interruption. The revision interpretation for Rev2+ will need to be handled appropriately in the future.

* Refactor SEG-Y revision parsing for clarity and correctness

Revised the `transform` method to enhance parsing of SEG-Y revisions from the binary header. Added detailed handling for SEG-Y Rev1 and simplified processing for Rev2 to ensure accurate interpretation based on the binary format.

* Clarify division base in SEGY revision calculation

Updated the calculation of the SEGY revision to specify that the division by 256.0 is based on hexadecimal representation. This enhances code readability and helps avoid potential confusion for future maintainers.

* Refactor SEGY revision handling in binary header

Standardize the logic for setting SEGY revisions and include minor version information for revisions above REV1. This ensures proper version encoding and enhances compatibility with higher SEGY standards.

* static typing fixess

* Refactor revision transform handling

Refactored how SEG-Y revisions are parsed and applied within the transform pipelines. Introduced better handling for integer-based Rev1 and separated handling for major and minor revisions in Rev2. Added comprehensive tests for different endianness and revision versions to ensure reliability.

* static typing for mypy

* Refactor segy_revision property and clean up tests

Refactored the segy_revision property to ensure a default value of SegyStandard.REV0 when the segy_standard is None. Removed redundant checks in the create_textual_header method and cleaned up type-ignore comments in the test cases. This improves readability and robustness of the SEG-Y factory code.

* ensure we use the float value of standard

* Add None option to segy_standard in factory test configs

Updated the SegyFactoryTestConfig class to allow segy_standard to be None. This change expands the test configurations to include cases where segy_standard is not specified.

---------

Co-authored-by: Altay Sansal <altay.sansal@tgs.com>